basic scaffolding for modeling a Tweet
Just the model file and the expected attributes for now.  This
will be used to parse and access Tweet data.

This brings in the `t` gem as a dependency.  The gem will be used
to fetch tweets and create objs from it.

This also includes some reorgs to some setup logic.