rework daemon setup to be run against a single `TWEETME_USER`
This updates the daemon setup logic to expect a single username in
the env var `TWEETME_USER`.  This user represents the configuration
the daemon will run against.  This implies you can only run one
user per daemon process.

This was a simpler approach for running multiple users/logs/etc than
trying to have a single daemon run multiple users.  This removes
some unnecessary logic from the logger and the user classes now
that we are moving to this design.

Finally, I moved the daemon setup file to the config dir b/c this is
a more appropriate location for the file as it is setup code, not
lib code.